requirements.txt
torch==2.4.0
torchvision
xformers
triton
runpod==1.7.7
diffusers==0.11.1
transformers==4.38.0
huggingface-hub==0.25.2
imageio==2.27.0
decord==0.6.0
accelerate==0.26.1
einops==0.7.0
omegaconf==2.3.0
safetensors==0.4.2
opencv-python==4.9.0.80
mediapipe==0.10.11
av==11.0.0
torch-fidelity==0.3.0
python_speech_features==0.6
librosa==0.10.1
scenedetect==0.6.1
ffmpeg-python==0.2.0
lpips==0.1.4
face-alignment==1.4.1
ninja==1.11.1.1
pandas==2.0.3
numpy==1.24.4
gradio==5.12.0
b2sdk
python-dotenv